+++ src/index.tsx
@@ -1,15 +1,9 @@
-import React, { useState } from "react";
+import React from "react";
 import ReactDOM from "react-dom";
+import './style/index.less';
 
-function Person() {
-  let [user, setUser] = useState("张三");
-  return (
-    <div>
-      <h1>今天{user}去七宝老街了，中午的人数不是很多，吃了葱油饼很油腻</h1>
-      <button onClick={()=> user === '张三' ? setUser('李四') : setUser('张三') }>换人去逛街</button>
-    </div>
-  );
-}
+import Person from './Person';
+
 
 ReactDOM.render(<Person />, document.getElementById("app")); //低于react v18
 // v18-react