+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## 1.0.0
 
-- 代码格式化
+- 代码格式化 npm init @eslint/config
 - less
 - react热更新
 - 强缓存优化
--
+- css 单独打包